Custom endpoint usage
